% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board-server.R
\name{block_panel_server}
\alias{block_panel_server}
\title{Block Panel Server}
\usage{
block_panel_server(board, update, session)
}
\arguments{
\item{board}{Board reactiveValues}

\item{update}{Update function}

\item{session}{Shiny session}
}
\description{
Handles the slide-out block panel for adding/appending blocks.
}
\keyword{internal}
