% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action-modal.R
\name{block_panel_ui}
\alias{block_panel_ui}
\title{Block Panel UI}
\usage{
block_panel_ui(ns)
}
\arguments{
\item{ns}{Namespace function}
}
\value{
Shiny UI element
}
\description{
Creates a slide-out panel for block selection with categorized blocks.
}
\keyword{internal}
