% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar-ui.R
\name{blockr_sidebar_dep}
\alias{blockr_sidebar_dep}
\title{Sidebar JS Dependency}
\usage{
blockr_sidebar_dep()
}
\value{
htmltools::htmlDependency object
}
\description{
Creates HTML dependency for the blockr sidebar JavaScript module.
}
\keyword{internal}
