% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar-ui.R
\name{code_button_ui}
\alias{code_button_ui}
\title{Code button UI}
\usage{
code_button_ui(ns)
}
\arguments{
\item{ns}{Namespace function}
}
\value{
Shiny tag
}
\description{
Button that triggers code generation.
}
\keyword{internal}
