% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{stack_color}
\alias{stack_color}
\alias{suggest_new_colors}
\title{Stack color}
\usage{
stack_color(x)

suggest_new_colors(colors = character(), n = 1)
}
\arguments{
\item{x}{Stack object}

\item{colors}{Currently used color values}

\item{n}{Number of new colors to generate}
}
\value{
A character vector of (hex encoded) color value(s).
}
\description{
While stacks created via \code{\link[blockr.core:new_stack]{blockr.core::new_stack()}} do not keep track of
a color attribute, there are stack objects that inherit from "stack" and
do so. Such objects may implement the generic \code{stack_color()} in order to
provide a mechanism to retrieve a color value.
}
