% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar-content.R
\name{default_block_icon}
\alias{default_block_icon}
\title{Default Block Icon}
\usage{
default_block_icon()
}
\value{
HTML string for default icon
}
\description{
Default Block Icon
}
\keyword{internal}
