% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-overview.R
\name{delete_confirm_modal}
\alias{delete_confirm_modal}
\title{Delete confirmation modal}
\usage{
delete_confirm_modal(ns, count)
}
\arguments{
\item{ns}{Namespace function}

\item{count}{Number of workflows to delete}
}
\description{
Delete confirmation modal
}
\keyword{internal}
