% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-id.R
\name{dock_id}
\alias{dock_id}
\alias{as_obj_id}
\alias{as_block_panel_id}
\alias{as_ext_panel_id}
\alias{as_block_handle_id}
\title{ID utilities}
\usage{
dock_id(ns = NULL)

as_obj_id(x)

as_block_panel_id(x)

as_ext_panel_id(x)

as_block_handle_id(x)
}
\arguments{
\item{ns}{Namespace prefix}

\item{x}{Object}
}
\description{
Functions for converting "object IDs" (i.e. block or extension IDs) to
panel IDs and other ID handling utilities.
}
