% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.R
\name{dock_remove_view}
\alias{dock_remove_view}
\title{Remove a view programmatically}
\usage{
dock_remove_view(name, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{name}{View name to remove.}

\item{session}{The Shiny session. Defaults to the current one.}
}
\value{
Invisibly, the new active view name.
}
\description{
Destroys the view's dock module and switches away if it was active.
Refuses to remove the last remaining view.
}
