% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.R
\name{dock_set_active_view}
\alias{dock_set_active_view}
\title{Switch the active view programmatically}
\usage{
dock_set_active_view(view_name, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{view_name}{Name of the view to activate.}

\item{session}{The Shiny session. Defaults to the current one.}
}
\value{
Invisibly, the new active view name.
}
\description{
Drives view switching from outside the dock UI events. Same effect
as the user clicking a view tab.
}
