% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext-edit.R
\name{new_edit_board_extension}
\alias{new_edit_board_extension}
\title{Edit board extension}
\usage{
new_edit_board_extension(...)
}
\arguments{
\item{...}{Forwarded to \code{new_dock_extension()}}
}
\value{
A board extension object that additionally inherits from
\code{edit_board_extension}.
}
\description{
A simplistic example of an extension which can be used for manipulating
the board via a table-based UI. Mainly relevant for testing purposes.
}
\examples{
ext <- new_edit_board_extension()
is_dock_extension(ext)

}
