% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar-content.R
\name{empty_sidebar_content}
\alias{empty_sidebar_content}
\title{Empty Sidebar Content}
\usage{
empty_sidebar_content(message)
}
\arguments{
\item{message}{The message to display}
}
\value{
Shiny tags
}
\description{
Renders a consistent empty state message for sidebars.
}
\keyword{internal}
