% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar-server.R
\name{format_time_ago}
\alias{format_time_ago}
\title{Format timestamp as relative time}
\usage{
format_time_ago(timestamp)
}
\arguments{
\item{timestamp}{POSIXct timestamp}
}
\value{
Character string like "2 min ago", "3 hours ago", etc.
}
\description{
Format timestamp as relative time
}
\keyword{internal}
