% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar-server.R
\name{get_initials}
\alias{get_initials}
\title{Get initials from a username}
\usage{
get_initials(username)
}
\arguments{
\item{username}{Character string}
}
\value{
Character string with 1-2 uppercase initials
}
\description{
Get initials from a username
}
\keyword{internal}
