% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar-ui.R
\name{get_navbar_left}
\alias{get_navbar_left}
\title{Get navbar left slot content}
\usage{
get_navbar_left(ns, board)
}
\arguments{
\item{ns}{Namespace function}

\item{board}{A dock_board object}
}
\value{
tagList of UI elements
}
\description{
Collects left slot UI from all navbar_provider objects.
}
\keyword{internal}
