% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar-api.R
\name{get_sidebar_context}
\alias{get_sidebar_context}
\title{Get Sidebar Context}
\usage{
get_sidebar_context(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{Shiny session}
}
\value{
List of context values
}
\description{
Retrieves context data set by \code{set_sidebar_context()}.
}
