% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar-api.R
\name{get_sidebar_type}
\alias{get_sidebar_type}
\title{Get Current Sidebar Type}
\usage{
get_sidebar_type(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{Shiny session}
}
\value{
Sidebar type object or NULL if sidebar is hidden
}
\description{
Returns the current sidebar type object (for S3 dispatch).
}
\keyword{internal}
