% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-block.R
\name{update_block_ui}
\alias{update_block_ui}
\title{Update some pieces of the block UI}
\usage{
update_block_ui(board, update, session, parent, ...)
}
\description{
Some elements of the block UI require server
elements to be available. This can't be done
from the board_ui as this one is done once.
}
\keyword{internal}
