% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar-server.R
\name{init_navbar_servers}
\alias{init_navbar_servers}
\title{Initialize navbar provider servers}
\usage{
init_navbar_servers(
  board,
  dock = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{board}{Reactive values object containing board state}

\item{dock}{Dock object (optional)}

\item{session}{Shiny session}
}
\description{
Calls server functions from all navbar_provider objects.
}
\keyword{internal}
