% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar-ui.R
\name{navbar_ui}
\alias{navbar_ui}
\title{Navbar UI Module}
\usage{
navbar_ui(id)
}
\arguments{
\item{id}{Namespace ID}
}
\value{
Shiny UI element
}
\description{
Creates the top navbar with workflow management controls.
}
\keyword{internal}
