% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dock-board.R
\name{new_dock_board}
\alias{new_dock_board}
\alias{is_dock_board}
\title{Dock board}
\usage{
new_dock_board(..., class = character())

is_dock_board(x)
}
\arguments{
\item{...}{Further (metadata) attributes}

\item{class}{Board sub-class}

\item{x}{Board object}
}
\value{
A \code{board} object
}
\description{
Using the docking layout manager provided by dockViewR, a \code{dock_board}
extends \code{\link[blockr.core:new_board]{blockr.core::new_board()}}.
}
