% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar-ui.R
\name{options_offcanvas_ui}
\alias{options_offcanvas_ui}
\title{Settings offcanvas UI}
\usage{
options_offcanvas_ui(id, board, options_ui, extensions_ui)
}
\arguments{
\item{id}{Namespace ID}

\item{board}{A dock_board object}

\item{options_ui}{UI for board options}

\item{extensions_ui}{UI for extension options}
}
\value{
Shiny tag
}
\description{
Offcanvas sidebar for board settings and extension options.
}
\keyword{internal}
