% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar-ui.R
\name{render_category_header}
\alias{render_category_header}
\title{Render Category Header}
\usage{
render_category_header(title)
}
\arguments{
\item{title}{Category title}
}
\value{
Shiny tags
}
\description{
Render Category Header
}
\keyword{internal}
