% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar-content.R
\name{render_stack_sidebar_footer}
\alias{render_stack_sidebar_footer}
\title{Render Stack Sidebar Footer}
\usage{
render_stack_sidebar_footer(ns, board, mode, stack = NULL, stack_id = NULL)
}
\arguments{
\item{ns}{Namespace function}

\item{board}{The board object}

\item{mode}{One of "create", "edit"}

\item{stack}{For edit mode, the stack object}

\item{stack_id}{For edit mode, the stack ID}
}
\value{
Shiny tags
}
\description{
Generates the footer for stack sidebars with collapsible options.
}
\keyword{internal}
