% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-overview.R
\name{render_workflow_table}
\alias{render_workflow_table}
\title{Render workflow table}
\usage{
render_workflow_table(workflows, ns, load_input_id, select_input_id)
}
\arguments{
\item{workflows}{Data frame of workflows}

\item{ns}{Namespace function}

\item{load_input_id}{Input ID for load action}

\item{select_input_id}{Input ID for selection changes}
}
\description{
Render workflow table
}
\keyword{internal}
