% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar-api.R
\name{set_sidebar_context}
\alias{set_sidebar_context}
\title{Set Sidebar Context}
\usage{
set_sidebar_context(..., session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{...}{Named values to store in context}

\item{session}{Shiny session}
}
\description{
Sets context data that sidebar content renderers can access.
Call this before \code{show_sidebar()} when the content needs additional data.
}
\examples{
\dontrun{
# Set context, then show sidebar
set_sidebar_context(source_block = "blk_1")
show_sidebar("add_link")
}

}
