% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar-ui.R
\name{settings_button_ui}
\alias{settings_button_ui}
\title{Settings button UI}
\usage{
settings_button_ui(ns)
}
\arguments{
\item{ns}{Namespace function}
}
\value{
Shiny tag
}
\description{
Button that opens the settings offcanvas sidebar.
}
\keyword{internal}
