% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar-ui.R
\name{sidebars_ui}
\alias{sidebars_ui}
\title{Sidebar Containers UI}
\usage{
sidebars_ui(ns, board, generate_code_ui = NULL)
}
\arguments{
\item{ns}{Namespace function}

\item{board}{The board object}

\item{generate_code_ui}{Optional UI for the generate code plugin}
}
\value{
Shiny tags containing all sidebar containers
}
\description{
Creates all sidebar containers for a board.
}
\keyword{internal}
