% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-overview.R
\name{workflow_overview_ui}
\alias{workflow_overview_ui}
\title{Workflow Overview Page}
\usage{
workflow_overview_ui(ns)
}
\arguments{
\item{ns}{Namespace function}
}
\description{
Full page showing all saved workflows in a reactable table.
}
\keyword{internal}
