% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{new_dock_board}
\alias{new_dock_board}
\alias{is_dock_board}
\alias{dock_board_options}
\title{Dock board}
\usage{
new_dock_board(..., options = dock_board_options(), class = character())

is_dock_board(x)

dock_board_options()
}
\arguments{
\item{...}{Further (metadata) attributes}

\item{options}{Board-level user settings}

\item{class}{Board sub-class}

\item{x}{Board object}
}
\value{
A \code{board} object
}
\description{
Using the docking layout manager provided by dockViewR, a \code{dock_board}
extends \code{\link[blockr.core:new_board]{blockr.core::new_board()}}.
}
