% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R, R/layout.R
\name{new_dock_board}
\alias{new_dock_board}
\alias{is_dock_board}
\alias{board_layout}
\alias{board_layout<-}
\alias{dock_board_options}
\alias{new_board_layout}
\alias{is_board_layout}
\alias{validate_board_layout}
\alias{as_board_layout}
\alias{layout_panel_block_ids}
\title{Dock board}
\usage{
new_dock_board(
  ...,
  layout = new_board_layout(),
  options = dock_board_options(),
  class = character()
)

is_dock_board(x)

board_layout(x)

board_layout(x) <- value

dock_board_options()

new_board_layout(grid = NULL, panels = NULL, active_group = NULL)

is_board_layout(x)

validate_board_layout(x, blocks = character())

as_board_layout(x, ...)

layout_panel_block_ids(x)
}
\arguments{
\item{...}{Further (metadata) attributes}

\item{layout}{Dock layout}

\item{options}{Board-level user settings}

\item{class}{Board sub-class}

\item{x}{Board object}

\item{value}{Replacement value}

\item{grid, panels, active_group}{Layout components}

\item{blocks}{Block IDs}
}
\value{
A \code{board} object
}
\description{
Using the docking layout manager provided by dockViewR, a \code{dock_board}
extends \code{\link[blockr.core:new_board]{blockr.core::new_board()}}.
}
