% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board-class.R, R/layout-class.R
\name{new_dock_board}
\alias{new_dock_board}
\alias{is_dock_board}
\alias{dock_layout}
\alias{dock_layout<-}
\alias{dock_extensions}
\alias{dock_extensions<-}
\alias{dock_ext_ids}
\alias{dock_board_options}
\alias{new_dock_layout}
\alias{default_dock_layout}
\alias{is_dock_layout}
\alias{validate_dock_layout}
\alias{as_dock_layout}
\alias{layout_panel_ids}
\title{Dock board}
\usage{
new_dock_board(
  blocks = list(),
  ...,
  extensions = list(),
  layout = default_dock_layout(blocks, extensions),
  options = dock_board_options(),
  class = character()
)

is_dock_board(x)

dock_layout(x)

dock_layout(x) <- value

dock_extensions(x)

dock_extensions(x) <- value

dock_ext_ids(x)

dock_board_options()

new_dock_layout(grid = NULL, panels = NULL, active_group = NULL)

default_dock_layout(blocks = list(), extensions = list())

is_dock_layout(x)

validate_dock_layout(x, blocks = character())

as_dock_layout(x, ...)

layout_panel_ids(x)
}
\arguments{
\item{blocks}{Block IDs}

\item{...}{Further (metadata) attributes}

\item{extensions}{Dock extensions}

\item{layout}{Dock layout}

\item{options}{Board-level user settings}

\item{class}{Board sub-class}

\item{x}{Board object}

\item{value}{Replacement value}

\item{grid, panels, active_group}{Layout components}
}
\value{
A \code{board} object
}
\description{
Using the docking layout manager provided by dockViewR, a \code{dock_board}
extends \code{\link[blockr.core:new_board]{blockr.core::new_board()}}.
}
