% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{new_dock_extension}
\alias{new_dock_extension}
\alias{is_dock_extenstion}
\title{Dock extensions}
\usage{
new_dock_extension(server, ui, class, ...)

is_dock_extenstion(x)
}
\arguments{
\item{server}{A function returning \code{\link[shiny:moduleServer]{shiny::moduleServer()}}}

\item{ui}{A function with a single argument (\code{ns}) returning a \code{shiny.tag}}

\item{class}{Extension subclass}
}
\description{
Functionality of a \code{dock_board} can be extended by supplying one or more
\code{dock_extension} objects, which essentially provide UI shown in a dock panel
that allows for manipulating the board state.
}
