% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext-class.R
\name{new_dock_extension}
\alias{new_dock_extension}
\alias{is_dock_extension}
\alias{validate_extension}
\alias{extension_ui}
\alias{extension_server}
\alias{extension_id}
\alias{extension_name}
\title{Dock extensions}
\usage{
new_dock_extension(server, ui, name, class, ...)

is_dock_extension(x)

validate_extension(x, ...)

extension_ui(x, id, ...)

extension_server(x, ...)

extension_id(x)

extension_name(x)
}
\arguments{
\item{server}{A function returning \code{\link[shiny:moduleServer]{shiny::moduleServer()}}}

\item{ui}{A function with a single argument (\code{ns}) returning a \code{shiny.tag}}

\item{name}{Name for extension}

\item{class}{Extension subclass}

\item{...}{Further attributes}

\item{x}{Extension object}

\item{id}{Namespace ID}
}
\description{
Functionality of a \code{dock_board} can be extended by supplying one or more
\code{dock_extension} objects, which essentially provide UI shown in a dock panel
that allows for manipulating the board state.
}
