% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-class.R
\name{new_dock_layout}
\alias{new_dock_layout}
\alias{create_dock_layout}
\alias{is_dock_layout}
\alias{validate_dock_layout}
\alias{as_dock_layout}
\alias{layout_panel_ids}
\title{Dock layout}
\usage{
new_dock_layout(grid = NULL, panels = NULL, active_group = NULL)

create_dock_layout(blocks = list(), extensions = list(), arrangement = NULL)

is_dock_layout(x)

validate_dock_layout(x, blocks = character())

as_dock_layout(x, ...)

layout_panel_ids(x)
}
\arguments{
\item{grid, panels, active_group}{Layout components}

\item{blocks, extensions}{Dock board components}

\item{arrangement}{Panel arrangement}

\item{x}{Object}

\item{...}{Generic consistency}
}
\description{
The arrangement of panels in a dock can be specified using a \code{dock_layout}
object. A default layout
}
