% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar-server.R
\name{navbar_server}
\alias{navbar_server}
\title{Navbar server}
\usage{
navbar_server(
  id,
  board,
  dock = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{Namespace ID}

\item{board}{Reactive values object containing board state}

\item{dock}{Dock object (optional, for serialization)}

\item{session}{Shiny session}
}
\description{
Initializes server logic for navbar, including all navbar_provider servers.
}
\keyword{internal}
