% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ui.R
\name{show_panel}
\alias{show_panel}
\title{UI utilities}
\usage{
show_panel(id, board, dock, type = c("block", "extension"))
}
\arguments{
\item{id}{Object ID}

\item{board}{Board object}

\item{dock}{Object available as \code{dock} in extensions}

\item{type}{Either "block" or "extensions", depending on what kind of panel
should be shown}
}
\value{
\code{NULL}, invisibly
}
\description{
Exported utilities for manipulating dock panels (i.e. displaying panels).
}
