% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar-api.R
\name{sidebar}
\alias{sidebar}
\alias{sidebar-api}
\alias{init_sidebar}
\alias{show_sidebar}
\alias{hide_sidebar}
\alias{get_sidebar}
\alias{new_sidebar}
\alias{pin_sidebar}
\alias{unpin_sidebar}
\alias{toggle_pin_sidebar}
\alias{is_sidebar_pinned}
\title{Sidebar API}
\usage{
init_sidebar(session = shiny::getDefaultReactiveDomain())

show_sidebar(id, context = list(), session = shiny::getDefaultReactiveDomain())

hide_sidebar(session = shiny::getDefaultReactiveDomain())

get_sidebar(session = shiny::getDefaultReactiveDomain())

new_sidebar(id, context = list())

pin_sidebar(session = shiny::getDefaultReactiveDomain())

unpin_sidebar(session = shiny::getDefaultReactiveDomain())

toggle_pin_sidebar(session = shiny::getDefaultReactiveDomain())

is_sidebar_pinned(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{Shiny session (default: current reactive domain)}

\item{id}{Either a string sidebar ID or a sidebar object from \code{\link[=new_sidebar]{new_sidebar()}}}

\item{context}{List of context data (ignored if \code{id} is a sidebar object)}
}
\description{
A slide-out sidebar system for blockr.dock. The API mirrors Shiny's
modal API:\tabular{ll}{
   Shiny Modal \tab Sidebar \cr
   \code{showModal(modalDialog(...))} \tab \code{show_sidebar(new_sidebar(...))} \cr
   \code{removeModal()} \tab \code{hide_sidebar()} \cr
}
}
\section{Functions}{

\itemize{
\item \code{\link[=init_sidebar]{init_sidebar()}} - Initialize the sidebar system (call once in server)
\item \code{\link[=show_sidebar]{show_sidebar()}} - Show a sidebar (like \code{\link[shiny:showModal]{shiny::showModal()}})
\item \code{\link[=hide_sidebar]{hide_sidebar()}} - Hide the current sidebar (like \code{\link[shiny:showModal]{shiny::removeModal()}})
\item \code{\link[=new_sidebar]{new_sidebar()}} - Create a sidebar object (like \code{\link[shiny:modalDialog]{shiny::modalDialog()}})
\item \code{\link[=get_sidebar]{get_sidebar()}} - Get current sidebar state
}
}

\section{Usage}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{# In server (once):
init_sidebar(session)

# Show a sidebar:
show_sidebar(new_sidebar("add_block", list(mode = "add")))

# Or with custom ID and context:
show_sidebar("my_sidebar", context = list(data = myData))

# Hide:
hide_sidebar()

# Get state (for conditional rendering):
sidebar <- get_sidebar()
if (identical(sidebar$sidebar_id, "my_sidebar")) \{
  # render content...
\}
}\if{html}{\out{</div>}}
}

\section{How It Works}{

\enumerate{
\item \code{init_sidebar()} creates reactive state in \code{session$userData}
\item \code{show_sidebar()} sets the \code{sidebar_id} to show
\item An observer toggles CSS classes via shinyjs to show/hide
\item UI containers are always in the DOM, just hidden via CSS transform
}
}

